requirements.txt
google-adk==1.0.0
yfinance==0.2.18
pandas==2.0.0
numpy==1.24.0
ta-lib==0.4.0
requests==2.28.0
textblob==0.17.1
plotly==5.13.0
kaleido==0.2.1
